requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
python-dotenv==0.20.0
google-generativeai==0.1.0
SQLAlchemy==1.4.36